Remove the engine update-config action
It was originally added in 2014 to update the config from an old format, and
the prompt to call it was removed in v0.9999 in 2019, when it was formally
deprecated. Should anyone need to go back and use it (which seems quite
unlikely), it will still be available via older verions.