Lock destination after registry creation
Turns out that v1.2.0 introduced a bug when trying to deploy to a MySQL server
that does not already have a registry database (usually named "sqitch"). It
would fail because it could not connect to the registry. Which makes sense,
since on first deploy it does not usually exist yet.

So move the calls to lock the destination to after we're certain the registry
exists. For the deploy method, that's after the plan has been evaluated and
the registry created if it does not yet exist. For the revert method, it's
after the registry has been checked.

Unfortunately this introduces a potential race condition where two instances
of Sqitch are trying to work at the same time, and both try to create the
registry (both see it isn't there, one creates it, the other fails), but that
seems like an unlikely situation.

Resolves #601.

Resolves #601.